+++ src/dart_snippets/architecture/base_view_model.ts
@@ -10,28 +10,27 @@ export class BaseViewModel extends Base {
 
     this._dartString = `import 'package:flutter/foundation.dart';
 import 'package:logger/logger.dart';
-
 import '../logger.dart';
 
 class BaseViewModel extends ChangeNotifier {
-    String _title;
-    bool _busy;
-    Logger log;
-    bool _isDisposed = false;
+  String _title;
+  bool _busy;
+  Logger log;
+  bool _isDisposed = false;
 
-    BaseViewModel({
+  BaseViewModel({
     bool busy = false,
     String title,
-    })  : _busy = busy,
-        _title = title {
+  })  : _busy = busy,
+      _title = title {
     log = getLogger(title ?? this.runtimeType.toString());
-    }
+  }
 
-    bool get busy => this._busy;
-    bool get isDisposed => this._isDisposed;
-    String get title => _title ?? this.runtimeType.toString();
+  bool get busy => this._busy;
+  bool get isDisposed => this._isDisposed;
+  String get title => _title ?? this.runtimeType.toString();
 
-    set busy(bool busy) {
+  set busy(bool busy) {
     log.i(
         'busy: '
         '$title is entering '
@@ -39,24 +38,24 @@ class BaseViewModel extends ChangeNotifier {
     );
     this._busy = busy;
     notifyListeners();
-    }
+  }
 
-    @override
-    void notifyListeners() {
+  @override
+  void notifyListeners() {
     if (!isDisposed) {
         super.notifyListeners();
     } else {
         log.w('notifyListeners: Notify listeners called after '
             '\${title ?? this.runtimeType.toString()} has been disposed');
     }
-    }
+  }
 
-    @override
-    void dispose() {
+  @override
+  void dispose() {
     log.i('dispose');
     _isDisposed = true;
     super.dispose();
-    }
+  }
 }`;
   }