+++ src/dart_snippets/architecture/main.ts
@@ -16,21 +16,21 @@ import 'package:provider/provider.dart';
 import 'views/home/home_view.dart';
 
 void main() async {
-    await LocatorInjector.setupLocator();
-    runApp(MainApplication());
+  await LocatorInjector.setupLocator();
+  runApp(MainApplication());
 }
 
 class MainApplication extends StatelessWidget {
-    @override
-    Widget build(BuildContext context) {
+  @override
+  Widget build(BuildContext context) {
     return MultiProvider(
-        providers: ProviderInjector.providers,
-        child: MaterialApp(
+      providers: ProviderInjector.providers,
+      child: MaterialApp(
         navigatorKey: locator<NavigatorService>().navigatorKey,
         home: HomeView(),
-        ),
+      ),
     );
-    }
+  }
 }`;
   }