+++ src/utils/view_file.ts
@@ -7,11 +7,12 @@ import { ViewModel } from '../dart_snippets/views/view_model';
 import { Mobile } from '../dart_snippets/views/mobile';
 import { Desktop } from '../dart_snippets/views/desktop';
 import { Tablet } from '../dart_snippets/views/tablet';
+import { YamlHelper } from './yaml_helper';
 
 export class ViewFile {
 
-    constructor(private rootPath: string, private fileName: string) {
-        console.log(`ViewFile(rootPath: ${rootPath}, fileName: ${fileName})`);
+    constructor(private rootPath: string, private fileName: string, private folders?: string[]) {
+        console.debug(`ViewFile(rootPath: ${rootPath}, fileName: ${fileName})`);
         let folderCreated = FileSystemManager.createFolder(this.pathValue);
         if (!folderCreated) { return; }
     }
@@ -22,6 +23,14 @@ export class ViewFile {
         this.createTablet();
         this.createDesktop();
         this.createWithViewModel();
+    }
+
+    public createDemoViews() {
+        this.createFiles(this.snakeCasedFileName + '_view.dart', new View(this.snakeCasedFileName, 'View').dartString);
+        this.createFiles(this.snakeCasedFileName + '_mobile.dart', new Mobile(this.snakeCasedFileName, 'Mobile').demoString);
+        this.createFiles(this.snakeCasedFileName + '_desktop.dart', new Desktop(this.snakeCasedFileName, 'Desktop').demoString);
+        this.createFiles(this.snakeCasedFileName + '_tablet.dart', new Tablet(this.snakeCasedFileName, 'Tablet').demoString);
+        this.createFiles(this.snakeCasedFileName + '_view_model.dart', new ViewModel(this.snakeCasedFileName, 'ViewModel', YamlHelper.getProjectName()).demoString);
     }
 
     public createView() {
@@ -41,22 +50,25 @@ export class ViewFile {
     }
 
     public createWithViewModel() {
-        this.createFiles(this.snakeCasedFileName + '_view_model.dart', new ViewModel(this.snakeCasedFileName, 'ViewModel').dartString);
+        this.createFiles(this.snakeCasedFileName + '_view_model.dart', new ViewModel(this.snakeCasedFileName, 'ViewModel', YamlHelper.getProjectName()).dartString);
     }
 
     private get snakeCasedFileName(): string {
         let snakeCasedFileName = _.snakeCase(this.fileName);
-        console.log(`get snakeCasedFileName: ${snakeCasedFileName}`);
+        console.debug(`get snakeCasedFileName: ${snakeCasedFileName}`);
         return snakeCasedFileName;
     }
 
     private get pathValue(): string {
-        return path.join(
-            this.rootPath,
-            'lib',
-            'views',
-            this.snakeCasedFileName,
-        );
+        if (this.folders === undefined) {
+            return path.join(
+                this.rootPath,
+                'lib',
+                'views',
+                this.snakeCasedFileName
+            );
+        }
+        return path.join(this.rootPath, 'lib', 'views', ...this.folders, this.snakeCasedFileName);
     }
 
     private createFiles(fileName: string, data: string) {