+++ CHANGELOG.md
@@ -5,8 +5,11 @@ All notable changes to the "flutter-mvvm-architecture-generator" extension will 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+- Create folders for the `Create Widget` command
 
-## [Unreleased]
+## [1.4.0]
+- Create Folder if `CreateViews` command is given a parameter like `demo/login`
 
 ## [1.0.0] - 2019-12-17
 - Initial release